% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{biquadrates}
\alias{biquadrates}
\title{biquadrates a number}
\usage{
biquadrates(x)
}
\arguments{
\item{x}{A numeric value to be biquadrates}
}
\value{
the biquadrates of input
}
\description{
Takes in any numeric value and biquadrates it.
}
