% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{BUS-class}
\alias{BUS}
\alias{BUS-class}
\title{BUS}
\description{
The class BUS
}
\section{Fields}{

\describe{
\item{\code{n}}{total number of buses}

\item{\code{index}}{bus indexes}

\item{\code{int}}{internal nus number}

\item{\code{Pg}}{active power injected in the network by generators}

\item{\code{Qg}}{reactive power injected in the network by generators}

\item{\code{Pl}}{active power absorbed from the network by loads}

\item{\code{Ql}}{reactive power absorbed from the network by loads}

\item{\code{names}}{bus names}
}}

