% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{DEMAND-class}
\alias{DEMAND}
\alias{DEMAND-class}
\title{The class DEMAND}
\description{
The class DEMAND
}
\section{Fields}{

\describe{
\item{\code{data}}{demand load data}

\item{\code{n}}{total number of demand loads}
}}

