% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{fifth_power}
\alias{fifth_power}
\title{fifth power a number}
\usage{
fifth_power(x)
}
\arguments{
\item{x}{A numeric value to be fifth power}
}
\value{
the fifth power of input
}
\description{
Takes in any numeric value and fifth power it.
}
