% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{LINE-class}
\alias{LINE}
\alias{LINE-class}
\title{The class LINE}
\description{
The class LINE
}
\section{Fields}{

\describe{
\item{\code{data}}{transmission line data}

\item{\code{n}}{total number of lines}

\item{\code{Y}}{admittance matrix of the network}

\item{\code{from}}{indexes of buses at which lines begin}

\item{\code{to}}{indexes of buses at which lines end}
}}

