% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powerCASCADE}
\alias{powerCASCADE}
\title{powerCASCADE}
\usage{
powerCASCADE(L = 0.5, n = 1000, D = 4e-04, Q = 4e-04, N = 10000)
}
\arguments{
\item{L}{load stress}

\item{n}{total number of motors connected to load bus}

\item{D}{size of initial disturbance of CASCADE model}

\item{Q}{the amount of propagation in CASCADE model}

\item{N}{number of iteration}
}
\description{
analyse and plot PDF for CASCADE failing model
}
\examples{
powerCASCADE()
}

