% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powerData}
\alias{powerData}
\title{powerData}
\usage{
powerData(data = "d_003.R", path = paste(path.package("powerr"), "/extdata",
  sep = ""))
}
\arguments{
\item{data}{name of data file with '.R' extension}

\item{path}{the path of the datafile with no '/' in the end}
}
\description{
Read formated data into powerr
}
\examples{
\donttest{
powerData('d_003.R')
}
}

