% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powerData2powerr}
\alias{powerData2powerr}
\title{powerData2powerr}
\usage{
powerData2powerr(dataName = "d_006_mdl.m",
  dataPath = paste(path.package("powerr"), "/extdata", sep = ""),
  fromMode = "psat")
}
\arguments{
\item{dataName}{name of data file to be changed format}

\item{dataPath}{path of the file with no '/' in the end}

\item{fromMode}{the former data format}
}
\description{
change data file to powerr .R file format
}
\examples{
powerData2powerr()
}

