% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powerDenseMatrix}
\alias{powerDenseMatrix}
\title{create dense matrix with row and column index}
\usage{
powerDenseMatrix(rowIndex, columnIndex, data, nrow, ncol)
}
\arguments{
\item{rowIndex}{specify the location of row of non-zero entries in the matrix}

\item{columnIndex}{specify the location of column of non-zero entries in the matrix}

\item{data}{non-zero entries on these certain locations}

\item{nrow}{number of row of the matrix}

\item{ncol}{number of column of the matrix}
}
\description{
create dense matrix with row and column index
}

