% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powerFind}
\alias{powerFind}
\title{find index of non 'num' in a given object}
\usage{
powerFind(input, num)
}
\arguments{
\item{input}{the objext needed to be manipulated}

\item{num}{the number needed to be avoided in finding}
}
\description{
find index of non 'num' in a given object
}

