% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powerIm}
\alias{powerIm}
\title{powerIm}
\usage{
powerIm(x)
}
\arguments{
\item{x}{sparse or dense complex matrix}
}
\value{
imagine part of the object
}
\description{
calculate imagine part of a complex matrix either sparse or dense
}
\examples{
# sparse complex
x <- list(sparseMatrix(1, 1, x = 1, dims = c(3,3)), sparseMatrix(2, 2, x = 2, dims = c(3,3)))
powerIm(x) # 3 x 3 sparse Matrix of class "dgCMatrix"
# dense complex
x <- matrix(1+1i, 3, 3)
powerIm(x)
}

