% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powerPF}
\alias{powerPF}
\title{powerPF}
\usage{
powerPF(method = "newton", tolerance = 1e-05, iterLimit = 20)
}
\arguments{
\item{method}{"newton" for standard Newton Raphson method}

\item{tolerance}{set minimum accuracy for the power flow run}

\item{iterLimit}{set maximum iteration for each run}
}
\description{
solve standard power flow with Newton Raphson method
}
\examples{
\donttest{
powerPF()
}
}

