% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{powerr-class}
\alias{powerr}
\alias{powerr-class}
\title{top class of package powerr}
\description{
The top class powerr
}
\section{Fields}{

\describe{
\item{\code{data}}{powerr data}

\item{\code{store}}{store temp for data}

\item{\code{ncol}}{number of the column of the data input}

\item{\code{u}}{connection status}
}}

