% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powerSparseProduct}
\alias{powerSparseProduct}
\title{powerSparseProduct}
\usage{
powerSparseProduct(mRe1, mIm1, mRe2, mIm2)
}
\arguments{
\item{mRe1}{real part of first sparse matrix}

\item{mIm1}{imagine part of first sparse matrix}

\item{mRe2}{real part of second sparse matrix}

\item{mIm2}{imagine part of second sparse matrix}
}
\description{
use C++ to calculate matrix product between sparse martix including complex numbers
}

