% Generated by roxygen2 (4.0.2): do not edit by hand
\name{powerUnique}
\alias{powerUnique}
\title{return index of the unique number at its first appearence}
\usage{
powerUnique(input)
}
\arguments{
\item{input}{the object needed to be manipulated}
}
\description{
return index of the unique number at its first appearence
}

