% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{PQ-class}
\alias{PQ}
\alias{PQ-class}
\title{The class PQ}
\description{
The class PQ
}
\section{Fields}{

\describe{
\item{\code{data}}{PQ generator data}

\item{\code{n}}{total number of PQ generators}

\item{\code{bus}}{numbers of buses to which PQ generators are connected}

\item{\code{gen}}{1 if it is a PQ generator, 0 otherwise.}
}}

