% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{PV-class}
\alias{PV}
\alias{PV-class}
\title{PV}
\description{
The class PV
}
\section{Fields}{

\describe{
\item{\code{n}}{total number of PV generators}

\item{\code{bus}}{numbers of buses to which PV generators are connected}

\item{\code{pq}}{internal PQ bus data (used when generator reactive power limits are
encountered)}
}}

