% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SHUNT-class}
\alias{SHUNT}
\alias{SHUNT-class}
\title{SHUNT}
\description{
The class SHUNT
}
\section{Fields}{

\describe{
\item{\code{data}}{shunt impedance data.}

\item{\code{bus}}{numbers of buses to which shunt are connected}

\item{\code{g}}{column vector of the conductances at each bus of the network}

\item{\code{b}}{column vector of the susceptances at each bus of the network}
}}

