% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{sixth_power}
\alias{sixth_power}
\title{sixth power a number}
\usage{
sixth_power(x)
}
\arguments{
\item{x}{A numeric value to be sixth power}
}
\value{
the sixth power of input
}
\description{
Takes in any numeric value and sixth power it.
}
