% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{square}
\alias{square}
\title{Square a number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
the square of input
}
\description{
Takes in any numeric value and square it.
}
