% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{LINE-class}
\alias{LINE}
\alias{LINE-class}
\title{LINE}
\description{
The class LINE
}
\section{Fields}{

\describe{
\item{\code{n}}{total number of lines}

\item{\code{from}}{indexes of buses at which lines begin}

\item{\code{to}}{indexes of buses at which lines end}
}}

