% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{PQ-class}
\alias{PQ}
\alias{PQ-class}
\title{PQ}
\description{
The class PQ
}
\section{Fields}{

\describe{
\item{\code{n}}{total number of PQ generators}

\item{\code{bus}}{numbers of buses to which PQ generators are connected}

\item{\code{gen}}{1 if it is a PQ generator, 0 otherwise.}

\item{\code{P0}}{initial active power (used with non-conventional loads)}

\item{\code{Q0}}{initial reactive power (used with non-conventional loads)}
}}

