% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SG-class}
\alias{SG}
\alias{SG-class}
\title{SG}
\description{
The class SG (Slack Generator)
}
\section{Fields}{

\describe{
\item{\code{n}}{total number of slack generators}

\item{\code{bus}}{indexes of buses to which slack generators are connected}

\item{\code{vbus}}{indexes of voltage buses of slack generators}

\item{\code{refbus}}{indexes of buses used as phase reference}
}}

